회원제 게시판 기능 완료
회원을 관리하는 기능은 부족하지만, 회원제 게시판 기능의 구현 완료되었다.

다음 할일
- MySQL로 DB를 전환해보자, HSQLDB는 천천히 해도 될 것 같고..
  MySQL 8.0 부터 추가된 CTE 기능으로 답변형 게시글 목록을 구성해야할 것 같은데...
  잘 할 수 있을지 모르겠다. 🎃

+++ chap19/pro19/src/main/java/org/mklinkj/taojwp/board/BoardController.java
@@ -2,12 +2,12 @@ package org.mklinkj.taojwp.board;
 
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import static org.mklinkj.taojwp.common.constant.Constants.HTML_CONTENT_TYPE;
+import static org.mklinkj.taojwp.common.constant.Constants.LOGIN_INFO_KEY_NAME;
 import static org.mklinkj.taojwp.common.constant.Constants.MEGA_BYTE;
 import static org.mklinkj.taojwp.common.constant.Constants.PAGE_NAVI_SIZE;
 import static org.mklinkj.taojwp.common.constant.Constants.PAGE_SIZE;
 import static org.mklinkj.taojwp.common.constant.Constants.UPLOAD_DIR;
 import static org.mklinkj.taojwp.common.constant.Constants.UPLOAD_TEMP_DIR;
-import static org.mklinkj.taojwp.common.constant.Constants.UTF_8_ENCODING;
 import static org.mklinkj.taojwp.common.constant.Constants.VIEW_ROOT_PATH_FORMAT;
 import static org.mklinkj.taojwp.common.util.CommonUtils.fileNameOnly;
 
@@ -28,8 +28,10 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
+import org.mklinkj.taojwp.board.exception.InvalidRequestException;
 import org.mklinkj.taojwp.common.domain.ModalMessage;
 import org.mklinkj.taojwp.common.servlet.AbstractHttpServlet;
+import org.mklinkj.taojwp.member.MemberVO;
 
 @Slf4j
 @WebServlet("/board/*")
@@ -88,6 +90,8 @@ public class BoardController extends AbstractHttpServlet {
         nextPage = CURRENT_VIEW_PATH_FORMAT.formatted("/articleForm.jsp");
 
       } else if (action.equals("/addArticle.do")) {
+        HttpSession session = request.getSession();
+        MemberVO loginMember = (MemberVO) session.getAttribute(LOGIN_INFO_KEY_NAME);
 
         Map<String, String> articleMap = upload(request);
         String title = articleMap.get("title");
@@ -97,7 +101,7 @@ public class BoardController extends AbstractHttpServlet {
         ArticleVO articleVO =
             ArticleVO.builder() //
                 .parentNo(0)
-                .id("hong")
+                .id(loginMember.getId())
                 .title(title)
                 .content(content)
                 .imageFileName(imageFileName)
@@ -130,8 +134,19 @@ public class BoardController extends AbstractHttpServlet {
         nextPage = CURRENT_VIEW_PATH_FORMAT.formatted("/viewArticle.jsp");
 
       } else if (action.equals("/modArticle.do")) {
+        HttpSession session = request.getSession();
+        MemberVO loginMember = (MemberVO) session.getAttribute(LOGIN_INFO_KEY_NAME);
+        Integer articleNo = Integer.parseInt(request.getParameter("articleNo"));
+        ArticleVO dbArticle = boardService.viewArticle(articleNo);
+
+        if (!dbArticle.getId().equals(loginMember.getId())) {
+          LOGGER.error(
+              "잘못된 수정 요청: 게시글 작성자 ID {}, 로그인 ID: {}", dbArticle.getId(), loginMember.getId());
+          throw new InvalidRequestException(
+              ModalMessage.builder().title("잘못된 수정 요청").content("자신의 글만 수정 가능합니다. 😅").build());
+        }
+
         Map<String, String> articleMap = upload(request);
-        int articleNo = Integer.parseInt(articleMap.get("articleNo"));
         String title = articleMap.get("title");
         String content = articleMap.get("content");
         String imageFileName = articleMap.get("imageFileName");
@@ -170,7 +185,16 @@ public class BoardController extends AbstractHttpServlet {
                 "redirect:%s/viewArticle.do?articleNo=%s", request.getServletPath(), articleNo);
 
       } else if (action.equals("/removeArticle.do")) {
+        HttpSession session = request.getSession();
+        MemberVO loginMember = (MemberVO) session.getAttribute(LOGIN_INFO_KEY_NAME);
+
         Integer articleNo = Integer.parseInt(request.getParameter("articleNo"));
+        ArticleVO dbArticle = boardService.viewArticle(articleNo);
+
+        if (!dbArticle.getId().equals(loginMember.getId())) {
+          throw new InvalidRequestException(
+              ModalMessage.builder().title("잘못된 삭제 요청").content("자신의 글만 삭제 가능합니다. 😅").build());
+        }
 
         List<Integer> removedArticleNoList = boardService.removeArticle(articleNo);
         for (int removedArticleNo : removedArticleNoList) {
@@ -195,6 +219,8 @@ public class BoardController extends AbstractHttpServlet {
       } else if (action.equals("/addReply.do")) {
         HttpSession session = request.getSession();
         int parentNo = (Integer) session.getAttribute("parentNo");
+
+        MemberVO loginMember = (MemberVO) session.getAttribute(LOGIN_INFO_KEY_NAME);
         session.removeAttribute("parentNo");
 
         Map<String, String> articleMap = upload(request);
@@ -205,7 +231,7 @@ public class BoardController extends AbstractHttpServlet {
         ArticleVO articleVO =
             ArticleVO.builder() //
                 .parentNo(parentNo)
-                .id("lee")
+                .id(loginMember.getId())
                 .title(title)
                 .content(content)
                 .imageFileName(imageFileName)
@@ -235,6 +261,11 @@ public class BoardController extends AbstractHttpServlet {
       }
 
       forwardOrRedirect(request, response, nextPage);
+    } catch (InvalidRequestException ire) {
+      setFlashAttribute(request, "msg", ire.getModalMessage());
+      nextPage = String.format("redirect:%s/listArticles.do", request.getServletPath());
+      forwardOrRedirect(request, response, nextPage);
+
     } catch (Exception e) {
       LOGGER.info("서블릿 오류: {}", e.getMessage());
       setFlashAttribute(
